pyproject.toml
[tool.poetry]
name = "rubberduck-droid"
version = "0.1.0"
description = "Desktop speech enable droid assistant. Uses AI to discuss with you and answer your questions."
authors = ["Tommi Laukkanen"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
pvporcupine = "2.2.0"
pillow = "^11.0.0"
adafruit-circuitpython-ssd1306 = "^2.12.18"
aiohttp = "^3.10.10"
azure-core = "^1.31.0"
loguru = "^0.7.2"
pyaudio = "^0.2.14"
pydantic = "^2.9.2"
scipy = "^1.14.1"
python-dotenv = "^1.0.1"
rpi-gpio = "^0.7.1"
adafruit-blinka = "^8.49.0"
pvrecorder = "^1.2.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
